Add feature 1 - Release 1: Core Data Fetch & Parsing (MVP)
