+++ src/main/java/org/example/WeatherResponse.java
@@ -1,0 +1,13 @@
+package org.example;
+
+public class WeatherResponse {
+    private double temp;
+
+    public double getTemp() {
+        return temp;
+    }
+
+    public void setTemp(double temp) {
+        this.temp = temp;
+    }
+}
